Move run_a_module_script() to GlobalScope
Moving `run_a_module_script()` to `GlobalScope` will allow us to execute
module scripts on `WorkerGlobalScope`.

Prerequisite for #40365.

---

Relates to #23308
